Move general-purpose tags from demoscene.templatetags to common.templatetags
